генерация моков в отдельном файле, хелперы в utils
