+++ src/view/event-new.js
@@ -20,9 +20,7 @@ export const createNewTripTemplate = (trip = {}) => {
       return ``;
     }
 
-
-    return `
-    <section class="event__section  event__section--offers">
+    return `<section class="event__section  event__section--offers">
     <h3 class="event__section-title  event__section-title--offers">Offers</h3>
     <div class="event__available-offers">
